fix: Поправлен код для запуска на 11 авалонии
