+++ src/Unlimotion.Android/Unlimotion.Android.csproj
@@ -22,6 +22,10 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<PackageReference Include="Xamarin.AndroidX.Core.SplashScreen" Version="1.0.1" />
+	</ItemGroup>
+
+	<ItemGroup>
 	  <ProjectReference Include="..\Unlimotion\Unlimotion.csproj" />
 	</ItemGroup>
 </Project>
