+++ src/Unlimotion.iOS/Unlimotion.iOS.csproj
@@ -1,22 +1,27 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net7.0-ios</TargetFramework>
-    <SupportedOSPlatformVersion>10.0</SupportedOSPlatformVersion>
-    <ProvisioningType>manual</ProvisioningType>
-    <Nullable>enable</Nullable>
-    <RuntimeIdentifier>iossimulator-x64</RuntimeIdentifier>
-    
-    <!-- These properties need to be set in order to run on a real iDevice -->
-    <!--<RuntimeIdentifier>ios-arm64</RuntimeIdentifier>-->
-    <!--<CodesignKey></CodesignKey>-->
-  </PropertyGroup>
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFramework>net7.0-ios</TargetFramework>
+		<SupportedOSPlatformVersion>13.0</SupportedOSPlatformVersion>
+		<ProvisioningType>manual</ProvisioningType>
+		<Nullable>enable</Nullable>
+		<UseInterpreter>True</UseInterpreter>
+		<RuntimeIdentifier>iossimulator-x64</RuntimeIdentifier>
 
-  <ItemGroup>
-    <PackageReference Include="Avalonia.iOS" Version="$(AvaloniaVersion)" />
-  </ItemGroup>
+		<!-- These properties need to be set in order to run on a real iDevice -->
+		<!--<RuntimeIdentifier>ios-arm64</RuntimeIdentifier>-->
+		<!--<CodesignKey></CodesignKey>-->
+	</PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Unlimotion\Unlimotion.csproj" />
-  </ItemGroup>
+	<PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
+		<CodesignKey>iPhone Developer</CodesignKey>
+	</PropertyGroup>
+
+	<ItemGroup>
+		<PackageReference Include="Avalonia.iOS" Version="$(AvaloniaVersion)" />
+	</ItemGroup>
+
+	<ItemGroup>
+		<ProjectReference Include="..\Unlimotion\Unlimotion.csproj" />
+	</ItemGroup>
 </Project>
